Add step to install b24-php-sdk from source
This commit introduces an additional step in the GitHub Actions workflow to ensure the bitrix24/b24-php-sdk package is installed directly from source. This change is executed with options to prefer source installation and disable cache usage, which aids in maintaining the latest version of the package when running unit tests.

Signed-off-by: mesilov <mesilov.maxim@gmail.com>